+++ src/tasks.c
@@ -31,7 +31,7 @@ static void handle_boost_control() {
   if ((config.sensors[SENSOR_MAP].processed_value <
        config.boost_control.threshhold_kpa)) {
     if (config.sensors[SENSOR_TPS].processed_value > 90.0f) {
-      duty = 1.0f;
+      duty = 100.0f;
     } else {
       duty = 0.0f;
     }
@@ -39,7 +39,7 @@ static void handle_boost_control() {
     duty = interpolate_table_oneaxis(config.boost_control.pwm_duty_vs_rpm,
                                      config.decoder.rpm);
   }
-  set_pwm(config.boost_control.pin, duty);
+  set_pwm(config.boost_control.pin, duty / 100.0f);
 }
 
 static void handle_idle_control() {}